+++ repos/core/src/pages/eis/utils.js
@@ -184,3 +184,13 @@ export const convertFieldNameToTitleCaseText = (str) => {
 
   return result.charAt(0).toUpperCase() + result.slice(1);
 }
+
+/*
+  Return date in format 'YYYY-MM-DD'
+ */
+export const getDateStringFromDateObject = (date) => {
+  const year = String(date.getFullYear()).padStart(4, '0')
+  const month = String(date.getMonth() + 1).padStart(2, '0')
+  const day = String(date.getDate()).padStart(2, '0')
+  return `${year}-${month}-${day}`
+}